+++ src/js/addons/gif_search/addon.js
@@ -3,32 +3,25 @@ import { VideoList } from "./VideoList";
 import { FormatChatMsgProcHook } from "../../overrides";
 import { SAVEKEY_PFX } from "../../storage";
 import { createChatButton, createCheckbox, createModal, doesModalExist } from "../../ui/ui";
-import { addTextToChatline, arrayRemove, escapeHTML, /*getLast,*/ sendMessage } from "../../utils";
+import { addTextToChatline, arrayRemove, escapeHTML, sendMessage } from "../../utils";
 
 import * as css from "./style.css";
 
 const id = "gif-modal";
 const key = "GIFSEARCH";
-const cache = [];
 
 const KEY_gifFavorites    	 = SAVEKEY_PFX + "gifFavs";
-//const KEY_gifChatAutoplay	 = SAVEKEY_PFX + "gifChatAutoplay";
-//const KEY_gifMenuAutoplay	 = SAVEKEY_PFX + "gifMenuAutoplay";
 const KEY_gifChatEnabled	 = SAVEKEY_PFX + "gifChatEnabled";
 const KEY_gifSearchResultNum = SAVEKEY_PFX + "gifSearchResultNum";
 const KEY_gifAutoSend		 = SAVEKEY_PFX + "gifAutoSend";
 
 SETTINGS.gifFavorites 		= getOrDefault(KEY_gifFavorites, []);
-//SETTINGS.gifChatAutoplay 	= getOrDefault(KEY_gifChatAutoplay, true);
-//SETTINGS.gifMenuAutoplay 	= getOrDefault(KEY_gifMenuAutoplay, true);
 SETTINGS.gifChatEnabled 	= getOrDefault(KEY_gifChatEnabled, true);
 SETTINGS.gifSearchResultNum = getOrDefault(KEY_gifSearchResultNum, 20);
 SETTINGS.gifAutoSend 		= getOrDefault(KEY_gifAutoSend, true);
 
 $(window).unload(function() {
 	setOpt(KEY_gifFavorites			, SETTINGS.gifFavorites);
-	//setOpt(KEY_gifChatAutoplay  	, SETTINGS.gifChatAutoplay);
-	//setOpt(KEY_gifMenuAutoplay    	, SETTINGS.gifMenuAutoplay);
 	setOpt(KEY_gifChatEnabled 		, SETTINGS.gifChatEnabled);
 	setOpt(KEY_gifSearchResultNum	, SETTINGS.gifSearchResultNum);
 	setOpt(KEY_gifAutoSend			, SETTINGS.gifAutoSend);
@@ -39,7 +32,6 @@ let last_repost = 0;
 
 window.CLIENT.Nexus.plugins.loaded["gif_search"] = true;
 
-//const reg_tenor = /^https\:\/\/media\.tenor\.com\/[\w\-]+\/[\w\-]+\.(webm|mp4)$/;
 const reg_tenor = /^https\:\/\/media\.tenor\.com\/[\w\-]+\/[\%\w\-]+\.(webp|gif)$/;
 
 function onFavClick(el, URL, icon) {
@@ -164,8 +156,6 @@ function buildGifSearch() {
 				const r = data.results;
 				let c = 0;
 
-				//console.log(r);
-
 				for (let i = 0; i < r.length; i++) {
 
 					const URL = r[i].media_formats["tinywebp"].url;
@@ -174,16 +164,13 @@ function buildGifSearch() {
 						class: "tenor-wrap"
 					}).append(
 						createImageEmbed(URL, "tenor-embed", false, ()=>{onStandardClick(URL)})
-						//createVideoEmbed(URL, "webm", SETTINGS.gifMenuAutoplay, "tenor-embed", ()=>{onStandardClick(URL)})
 					).append(createFavButton(URL))
 						.appendTo(columns[c])
 
 					if (++c >= COLUMN_NUM) c = 0;
 
 				}
-			}/* else if (!ok && SETTINGS.DEBUG) {
-				window.debuglog(data);
-			}*/
+			}
 
 		})
 	}
@@ -248,61 +235,16 @@ function buildGifSearch() {
 			'</label>'+
 		'</div></div>'+
 	'</div>').insertAfter(inputgrp);
-		/*<div style="float: left;" class="checkbox input-sm">\
-    		<label for="gf-autoplay" title="If unchecked, videos will only play when hovered. Takes effect on your next search."><input type="checkbox" id="gf-autoplay"> Autoplay Results </label>\
-    	*/
 
 	$("#gf-in-resultnum").val(SETTINGS.gifSearchResultNum).on("change", function() {
 		setOpt(KEY_gifSearchResultNum, this.value);
 	})
-
-	/*$("#gf-autoplay").prop("checked", SETTINGS.gifMenuAutoplay).on("change", function(e) {
-		if (!e.originalEvent.isTrusted) return;
-		$("#gf-c-menuautoplay").prop("checked", this.checked);
-		SETTINGS.gifMenuAutoplay = this.checked;
-	});*/
-
-	/*$("#gf-autosend").prop("checked", SETTINGS.gifAutoSend).on("change", function(e) {
-		SETTINGS.gifAutoSend = this.checked;
-		setOpt(KEY_gifAutoSend, this.checked)
-	});*/
-/*
-	$('<nav aria-label="Page navigation" style="display: flex;width: 100%;">\
-		<ul class="pagination" style="margin: 10px auto;">\
-			<li>\
-			<a href="#" aria-label="Previous">\
-				<span aria-hidden="true">«</span>\
-			</a>\
-			</li>\
-			<li><a href="#">1</a></li>\
-			<li><a href="#">2</a></li>\
-			<li><a href="#">3</a></li>\
-			<li><a href="#">4</a></li>\
-			<li><a href="#">5</a></li>\
-			<li>\
-			<a href="#" aria-label="Next">\
-				<span aria-hidden="true">»</span>\
-			</a>\
-			</li>\
-		</ul>\
-	</nav>').insertBefore(container);*/
-
-	/*
-		<div style="float: left;" class="checkbox input-sm">\
-			<label for="gf-autosend" title="If checked, videos will be instantly sent in chat when clicked, instead of being sent to your chat input box."><input type="checkbox" id="gf-autosend"> Instantly post GIFs when clicked </label>\
-		</div>\*/
 
 	settings_body.append(createCheckbox("gf-c-enable", "Enable Search & Embeds", "gifChatEnabled", function(e, key) {
 		if (!e.originalEvent.isTrusted) return;
 		onCheckboxChange.call(this, e, key);
 		clearbtn.trigger("click");
 	}))
-	//.append(createCheckbox("gf-c-chatautoplay", "Autoplay Chat Embeds", "gifChatAutoplay", onCheckboxChange))
-	/*.append(createCheckbox("gf-c-menuautoplay", "Autoplay Menu Embeds", "gifMenuAutoplay", function(e, key) {
-		if (!e.originalEvent.isTrusted) return;
-		onCheckboxChange.call(this, e, key);
-		$("#gf-autoplay").prop("checked", this.checked);
-	}))*/
 	.append(createCheckbox("gf-c-autosend", "Instantly send GIFs from the menu when clicked", "gifAutoSend", onCheckboxChange))
 }
 
@@ -374,52 +316,6 @@ export function createImageEmbed(url, classes, showRepostBtn, onclick) {
 	return div.append(createFavButton(url)).append(createRepostButton(url));
 }
 
-/*export function createVideoEmbed(url, type, autoplay, vidclasses, onclick) {
-	if (!type) {
-		const ext = getLast(url.split("."));
-		if (ext == "webm") type = "webm";
-		else if (ext == "mp4") type = "mp4";
-		else throw new Error("createVideoEmbed: unknown extension (" + ext + "), only webm and mp4 accepted");
-	}
-
-	const video = $("<video/>", { muted: true, autoplay: false, loop: true, controlslist: "nofullscreen noremoteplayback", playsinline: true, disablepictureinpicture: true, disableremoteplayback: true })
-				.append($("<source/>", { src: escapeHTML(url), type: "video/" + type }));
-
-
-	//mute the damn thing. why do i have to explicitly mute it here anyway?
-	video[0].muted = true;
-
-	video.one("loadeddata", function() {
-		this.muted = true;
-		this.autoplay = autoplay;
-		if (this.autoplay) {
-			this.currentTime = 0.0;
-			this.play();
-		}
-	});
-
-	if (!autoplay) {
-		video.on("mouseenter", function() {
-			this.currentTime = 0.0;
-			this.play();
-		})
-		video.on("mouseleave", function() {
-			this.pause();
-			this.currentTime = 0.0;
-		})
-	}
-
-	if (vidclasses && typeof(vidclasses) == 'string') {
-		video.addClass(vidclasses);
-	}
-
-	if (onclick && typeof(onclick) == 'function') {
-		video.on("click", onclick);
-	}
-
-	return $("<div/>").append(video).append(createFavButton(url));
-}*/
-
 function processMessage(message, data) {
 	let found_tenor = false;
 	message.find("img").each(function() {
@@ -428,7 +324,6 @@ function processMessage(message, data) {
 		if (match) {
 			found_tenor = true;
 			if (SETTINGS.gifChatEnabled) {
-				//const embed = createVideoEmbed(this.href, match[1].toLowerCase(), SETTINGS.gifChatAutoplay);
 				const embed = createImageEmbed(this.src, "tenor-chat-embed", true);
 				$(this).replaceWith(embed);
 			} else {
@@ -436,18 +331,6 @@ function processMessage(message, data) {
 			}
 		}
 	})
-		/*
-		message.find("a").each(function() {
-			if (found_tenor) return;
-			const match = this.href.match(reg_tenor);
-			if (match) {
-				found_tenor = true;
-				const embed = createVideoEmbed(this.href, match[1].toLowerCase(), SETTINGS.gifChatAutoplay);
-				embed.addClass("tenor-chat-embed");
-				$(this).replaceWith(embed);
-			}
-			///^https\:\/\/media\.tenor\.com\/[\w-]+\/.+?\.(webm|mp4)$/
-		})*/
 	return message;
 }