+++ src/js/classes/Theme.js
@@ -10,9 +10,36 @@ export class Theme {
 		this.display_image = display_image;
 
 		themes[this.id] = this;
+
+		this.createOption();
 	}
 
-	apply() {
+	createOption() {
+
+		const self = this;
+
+		$("#opt-customthemes ." + this.id).remove();
+		const p = $("<p/>", {
+			class: this.id + " theme_option",
+			text: this.display_name,
+			title: this.id
+		}).on("click", function() {
+			this.blur();
+			self.apply(true);
+			$("#theme-success-text").text("Applied " + self.display_name + "!");
+		});
+
+		if (this.author && this.author.trim() != "") {
+			$("<span/>", {
+				text: "— created by: " + this.author,
+				class: "theme-author"
+			}).appendTo(p);
+		}
+
+		$("#opt-customthemes .theme-selection").append(p);
+	}
+
+	apply(saveTheme) {
 		const css_element = document.getElementsByClassName("custom-theme-css")[0];
 		if (this.cssIsLink) {
 			css_element.innerHTML = "@import url('"+this.css+"');"
@@ -28,5 +55,8 @@ export class Theme {
 				text: "active"
 			}).appendTo(option_element[0]);
 		}
+		
+		if (saveTheme)
+			SETTINGS.theme = this.id;
 	}
 }