+++ src/js/enum.js
@@ -1,4 +1,6 @@
 export const ChatAvatarSize = ["none","small","large"];
+export const ChatHeaderSize = ["off","small","normal"];
+export const ChatFooterSize = ["minimal","compact","normal"];
 export const NavbarSize = ["tiny","normal"];
 export const ChatSize = ["small","medium","large"];
 export const ChatWidthSize = ["tiny", "small", "normal", "large", "larger", "huge"];